+++ lib/clixon/clixon_xpath.h
@@ -150,5 +150,6 @@ int    xpath_vec(cxobj *xcur, cvec *nsc, const char *xpformat, cxobj ***vec, siz
 
 int xpath2canonical(const char *xpath0, cvec *nsc0, yang_stmt *yspec, char **xpath1, cvec **nsc1, cbuf **cbreason);
 int xpath_count(cxobj *xcur, cvec *nsc, const char *xpath, uint32_t *count);
+int xml2xpath(cxobj *x, cvec *nsc, int spec, char **xpath);
 
 #endif /* _CLIXON_XPATH_H */