+++ lib/clixon/clixon_yang_parse_lib.h
@@ -53,7 +53,7 @@
  */
 yang_stmt *yang_parse_file(FILE *fp, const char *name, yang_stmt *ysp);
 int        yang_file_find_match(clicon_handle h, const char *module, const char *revision, cbuf *fbuf);
-yang_stmt *yang_parse_filename(const char *filename, yang_stmt  *ysp);
+yang_stmt *yang_parse_filename(clicon_handle h, const char *filename, yang_stmt  *ysp);
 yang_stmt *yang_parse_module(clicon_handle h, const char *module, const char *revision, yang_stmt *yspec, char *origname);
 int        yang_parse_post(clicon_handle h, yang_stmt *yspec, int modmin);
 int        yang_spec_parse_module(clicon_handle h, const char *module,